Backend/requirements.txt
# Web 框架和工具
Flask==3.0.3
Flask-CORS==5.0.0
Werkzeug==3.0.6
python-dotenv==1.0.1

# 数据库
SQLAlchemy==2.0.36
PyMySQL==1.1.1

# 认证和安全
bcrypt==4.2.1
PyJWT==2.10.1

# 爬虫相关
selenium==4.27.1
beautifulsoup4==4.12.3
requests==2.32.3
webdriver-manager==4.0.2

# 任务调度
APScheduler==3.11.0

# 工具包
Pillow==11.0.0
